simple_rag_project/requirements.txt

streamlit
pandas
pypdf
python-docx
